Conflicts:
	bin/cassandra.in.bat
	bin/cassandra.ps1
	bin/cqlsh.bat
	bin/nodetool.bat
	bin/sstablelister.bat
	bin/sstableverify.bat
	bin/stop-server.ps1
	src/java/org/apache/cassandra/config/CFMetaData.java
	src/java/org/apache/cassandra/config/ColumnDefinition.java
	src/java/org/apache/cassandra/config/DatabaseDescriptor.java
	src/java/org/apache/cassandra/config/KSMetaData.java
	src/java/org/apache/cassandra/config/Schema.java
	src/java/org/apache/cassandra/config/UTMetaData.java
	src/java/org/apache/cassandra/cql3/ColumnIdentifier.java
	src/java/org/apache/cassandra/cql3/Constants.java
	src/java/org/apache/cassandra/cql3/Operation.java
	src/java/org/apache/cassandra/cql3/QueryProcessor.java
	src/java/org/apache/cassandra/cql3/UpdateParameters.java
	src/java/org/apache/cassandra/cql3/selection/Selection.java
	src/java/org/apache/cassandra/cql3/statements/AlterKeyspaceStatement.java
	src/java/org/apache/cassandra/cql3/statements/AlterTableStatement.java
	src/java/org/apache/cassandra/cql3/statements/CFPropDefs.java
	src/java/org/apache/cassandra/cql3/statements/CreateKeyspaceStatement.java
	src/java/org/apache/cassandra/cql3/statements/CreateTableStatement.java
	src/java/org/apache/cassandra/cql3/statements/DeleteStatement.java
	src/java/org/apache/cassandra/cql3/statements/KSPropDefs.java
	src/java/org/apache/cassandra/cql3/statements/ModificationStatement.java
	src/java/org/apache/cassandra/cql3/statements/SelectStatement.java
	src/java/org/apache/cassandra/cql3/statements/UpdateStatement.java
	src/java/org/apache/cassandra/db/AbstractCell.java
	src/java/org/apache/cassandra/db/AbstractRangeCommand.java
	src/java/org/apache/cassandra/db/ArrayBackedSortedColumns.java
	src/java/org/apache/cassandra/db/BufferCell.java
	src/java/org/apache/cassandra/db/BufferDeletedCell.java
	src/java/org/apache/cassandra/db/CollationController.java
	src/java/org/apache/cassandra/db/ColumnFamily.java
	src/java/org/apache/cassandra/db/ColumnFamilyStore.java
	src/java/org/apache/cassandra/db/ColumnIndex.java
	src/java/org/apache/cassandra/db/ColumnSerializer.java
	src/java/org/apache/cassandra/db/Directories.java
	src/java/org/apache/cassandra/db/ExpiringCell.java
	src/java/org/apache/cassandra/db/IndexExpression.java
	src/java/org/apache/cassandra/db/Memtable.java
	src/java/org/apache/cassandra/db/Mutation.java
	src/java/org/apache/cassandra/db/RangeTombstone.java
	src/java/org/apache/cassandra/db/ReadCommand.java
	src/java/org/apache/cassandra/db/Row.java
	src/java/org/apache/cassandra/db/SliceByNamesReadCommand.java
	src/java/org/apache/cassandra/db/SliceFromReadCommand.java
	src/java/org/apache/cassandra/db/SystemKeyspace.java
	src/java/org/apache/cassandra/db/TypeSizes.java
	src/java/org/apache/cassandra/db/columniterator/OnDiskAtomIterator.java
	src/java/org/apache/cassandra/db/commitlog/AbstractCommitLogService.java
	src/java/org/apache/cassandra/db/commitlog/CommitLog.java
	src/java/org/apache/cassandra/db/commitlog/CommitLogArchiver.java
	src/java/org/apache/cassandra/db/compaction/CompactionManager.java
	src/java/org/apache/cassandra/db/compaction/LeveledManifest.java
	src/java/org/apache/cassandra/db/compaction/SizeTieredCompactionStrategy.java
	src/java/org/apache/cassandra/db/composites/AbstractComposite.java
	src/java/org/apache/cassandra/db/composites/CellNameType.java
	src/java/org/apache/cassandra/db/composites/Composite.java
	src/java/org/apache/cassandra/db/composites/CompoundDenseCellNameType.java
	src/java/org/apache/cassandra/db/composites/CompoundSparseCellNameType.java
	src/java/org/apache/cassandra/db/filter/ColumnCounter.java
	src/java/org/apache/cassandra/db/filter/ColumnSlice.java
	src/java/org/apache/cassandra/db/filter/ExtendedFilter.java
	src/java/org/apache/cassandra/db/filter/NamesQueryFilter.java
	src/java/org/apache/cassandra/db/filter/QueryFilter.java
	src/java/org/apache/cassandra/db/filter/SliceQueryFilter.java
	src/java/org/apache/cassandra/db/index/AbstractSimplePerColumnSecondaryIndex.java
	src/java/org/apache/cassandra/db/index/PerColumnSecondaryIndex.java
	src/java/org/apache/cassandra/db/index/SecondaryIndex.java
	src/java/org/apache/cassandra/db/index/SecondaryIndexManager.java
	src/java/org/apache/cassandra/db/index/composites/CompositesIndexOnClusteringKey.java
	src/java/org/apache/cassandra/dht/BootStrapper.java
	src/java/org/apache/cassandra/dht/Murmur3Partitioner.java
	src/java/org/apache/cassandra/dht/Token.java
	src/java/org/apache/cassandra/io/compress/CompressedSequentialWriter.java
	src/java/org/apache/cassandra/io/compress/CompressionParameters.java
	src/java/org/apache/cassandra/io/sstable/Descriptor.java
	src/java/org/apache/cassandra/io/sstable/format/SSTableReader.java
	src/java/org/apache/cassandra/io/sstable/format/SSTableWriter.java
	src/java/org/apache/cassandra/io/sstable/format/big/BigTableScanner.java
	src/java/org/apache/cassandra/io/sstable/format/big/BigTableWriter.java
	src/java/org/apache/cassandra/io/sstable/format/big/SSTableSliceIterator.java
	src/java/org/apache/cassandra/io/sstable/format/big/SimpleSliceReader.java
	src/java/org/apache/cassandra/io/util/AbstractDataInput.java
	src/java/org/apache/cassandra/io/util/BufferedPoolingSegmentedFile.java
	src/java/org/apache/cassandra/io/util/CompressedPoolingSegmentedFile.java
	src/java/org/apache/cassandra/io/util/SequentialWriter.java
	src/java/org/apache/cassandra/net/IncomingTcpConnection.java
	src/java/org/apache/cassandra/net/OutboundTcpConnection.java
	src/java/org/apache/cassandra/schema/LegacySchemaTables.java
	src/java/org/apache/cassandra/service/AbstractReadExecutor.java
	src/java/org/apache/cassandra/service/CassandraDaemon.java
	src/java/org/apache/cassandra/service/MigrationManager.java
	src/java/org/apache/cassandra/service/ReadCallback.java
	src/java/org/apache/cassandra/service/RowDataResolver.java
	src/java/org/apache/cassandra/service/StorageProxy.java
	src/java/org/apache/cassandra/service/pager/MultiPartitionPager.java
	src/java/org/apache/cassandra/service/pager/QueryPagers.java
	src/java/org/apache/cassandra/service/pager/SinglePartitionPager.java
	src/java/org/apache/cassandra/service/paxos/PaxosState.java
	src/java/org/apache/cassandra/transport/Message.java
	src/java/org/apache/cassandra/transport/messages/AuthSuccess.java
	src/java/org/apache/cassandra/utils/CLibrary.java
	src/java/org/apache/cassandra/utils/IndexedSearchIterator.java
	test/unit/org/apache/cassandra/cql3/CreateTableTest.java~9274197b4bdb343731b964f2fcd8f70814f42a41

+++ src/java/org/apache/cassandra/cql3/statements/DropTypeStatement.java
@@ -21,9 +21,9 @@ import org.apache.cassandra.auth.Permission;
 import org.apache.cassandra.config.*;
 import org.apache.cassandra.cql3.*;
 import org.apache.cassandra.cql3.functions.Function;
-import org.apache.cassandra.cql3.functions.Functions;
 import org.apache.cassandra.db.marshal.*;
 import org.apache.cassandra.exceptions.*;
+import org.apache.cassandra.schema.KeyspaceMetadata;
 import org.apache.cassandra.service.ClientState;
 import org.apache.cassandra.service.MigrationManager;
 import org.apache.cassandra.transport.Event;
@@ -54,12 +54,11 @@ public class DropTypeStatement extends SchemaAlteringStatement
 
     public void validate(ClientState state) throws RequestValidationException
     {
-        KSMetaData ksm = Schema.instance.getKSMetaData(name.getKeyspace());
+        KeyspaceMetadata ksm = Schema.instance.getKSMetaData(name.getKeyspace());
         if (ksm == null)
             throw new InvalidRequestException(String.format("Cannot drop type in unknown keyspace %s", name.getKeyspace()));
 
-        UserType old = ksm.userTypes.getType(name.getUserTypeName());
-        if (old == null)
+        if (!ksm.types.get(name.getUserTypeName()).isPresent())
         {
             if (ifExists)
                 return;
@@ -74,30 +73,24 @@ public class DropTypeStatement extends SchemaAlteringStatement
         // we drop and 2) existing tables referencing the type (maybe in a nested
         // way).
 
-        for (Function function : Functions.all())
+        for (Function function : ksm.functions)
         {
             if (isUsedBy(function.returnType()))
                 throw new InvalidRequestException(String.format("Cannot drop user type %s as it is still used by function %s", name, function));
+
             for (AbstractType<?> argType : function.argTypes())
                 if (isUsedBy(argType))
                     throw new InvalidRequestException(String.format("Cannot drop user type %s as it is still used by function %s", name, function));
         }
 
-        for (KSMetaData ksm2 : Schema.instance.getKeyspaceDefinitions())
-        {
-            for (UserType ut : ksm2.userTypes.getAllTypes().values())
-            {
-                if (ut.keyspace.equals(name.getKeyspace()) && ut.name.equals(name.getUserTypeName()))
-                    continue;
-                if (isUsedBy(ut))
-                    throw new InvalidRequestException(String.format("Cannot drop user type %s as it is still used by user type %s", name, ut.asCQL3Type()));
-            }
+        for (UserType ut : ksm.types)
+            if (!ut.name.equals(name.getUserTypeName()) && isUsedBy(ut))
+                throw new InvalidRequestException(String.format("Cannot drop user type %s as it is still used by user type %s", name, ut.asCQL3Type()));
 
-            for (CFMetaData cfm : ksm2.cfMetaData().values())
-                for (ColumnDefinition def : cfm.allColumns())
-                    if (isUsedBy(def.type))
-                        throw new InvalidRequestException(String.format("Cannot drop user type %s as it is still used by table %s.%s", name, cfm.ksName, cfm.cfName));
-        }
+        for (CFMetaData cfm : ksm.tables)
+            for (ColumnDefinition def : cfm.allColumns())
+                if (isUsedBy(def.type))
+                    throw new InvalidRequestException(String.format("Cannot drop user type %s as it is still used by table %s.%s", name, cfm.ksName, cfm.cfName));
     }
 
     private boolean isUsedBy(AbstractType<?> toCheck) throws RequestValidationException
@@ -117,12 +110,6 @@ public class DropTypeStatement extends SchemaAlteringStatement
             CompositeType ct = (CompositeType)toCheck;
             for (AbstractType<?> subtype : ct.types)
                 if (isUsedBy(subtype))
-                    return true;
-        }
-        else if (toCheck instanceof ColumnToCollectionType)
-        {
-            for (CollectionType collection : ((ColumnToCollectionType)toCheck).defined.values())
-                if (isUsedBy(collection))
                     return true;
         }
         else if (toCheck instanceof CollectionType)
@@ -150,10 +137,10 @@ public class DropTypeStatement extends SchemaAlteringStatement
 
     public boolean announceMigration(boolean isLocalOnly) throws InvalidRequestException, ConfigurationException
     {
-        KSMetaData ksm = Schema.instance.getKSMetaData(name.getKeyspace());
+        KeyspaceMetadata ksm = Schema.instance.getKSMetaData(name.getKeyspace());
         assert ksm != null;
 
-        UserType toDrop = ksm.userTypes.getType(name.getUserTypeName());
+        UserType toDrop = ksm.types.getNullable(name.getUserTypeName());
         // Can be null with ifExists
         if (toDrop == null)
             return false;
